requirements.txt
Flask==3.0.3
openai==1.58.1
python-dotenv==0.21.0
